+++ kevlar/__init__.py
@@ -35,7 +35,6 @@ from kevlar.timer import Timer
 from kevlar import dump
 from kevlar import novel
 from kevlar import filter
-from kevlar import reaugment
 from kevlar import augment
 from kevlar import mutate
 from kevlar import assemble