+++ kevlar/augment.py
@@ -13,16 +13,23 @@ import screed
 import kevlar
 
 
-def augment(augreadstream, nakedseqtream):
+def augment(augseqstream, nakedseqstream):
+    """
+    Augment an unannotated stream of sequences.
+
+    - `augseqstream`: a stream of sequences annotated with k-mers of interest
+    - `nakedseqstream`: a stream of unannotated sequences, to be augmented with
+      k-mers of interest from `augseqstream`
+    """
     ksize = None
     ikmers = dict()
-    for record in augreadstream:
+    for record in augseqstream:
         for ikmer in record.ikmers:
             ikmers[ikmer.sequence] = ikmer.abund
             ikmers[kevlar.revcom(ikmer.sequence)] = ikmer.abund
             ksize = len(ikmer.sequence)
 
-    for record in nakedseqtream:
+    for record in nakedseqstream:
         newikmers = list()
         numkmers = len(record.sequence) - ksize + 1
         for offset in range(numkmers):
@@ -30,15 +37,22 @@ def augment(augreadstream, nakedseqtream):
             if kmer in ikmers:
                 ikmer = kevlar.KmerOfInterest(kmer, offset, ikmers[kmer])
                 newikmers.append(ikmer)
-        newrecord = screed.Record(name=record.name, sequence=record.sequence,
-                                  ikmers=newikmers)
+        if hasattr(record, 'quality'):
+            newrecord = screed.Record(
+                name=record.name, sequence=record.sequence, ikmers=newikmers,
+                quality=record.quality
+            )
+        else:
+            newrecord = screed.Record(
+                name=record.name, sequence=record.sequence, ikmers=newikmers
+            )
         yield newrecord
 
 
 def main(args):
     augfh = kevlar.open(args.augfastq, 'r')
-    augreads = kevlar.parse_augmented_fastx(augfh)
-    nakedreads = screed.open(args.sequences)
+    augseqs = kevlar.parse_augmented_fastx(augfh)
+    nakedseqs = screed.open(args.sequences)
     outstream = kevlar.open(args.out, 'w')
-    for record in augment(augreads, nakedreads):
+    for record in augment(augseqs, nakedseqs):
         kevlar.print_augmented_fastx(record, outstream)