+++ kevlar/cli/__init__.py
@@ -16,7 +16,6 @@ from . import count
 from . import effcount
 from . import novel
 from . import filter
-from . import reaugment
 from . import augment
 from . import assemble
 from . import mutate
@@ -33,7 +32,6 @@ mains = {
     'effcount': kevlar.effcount.main,
     'novel': kevlar.novel.main,
     'filter': kevlar.filter.main,
-    'reaugment': kevlar.reaugment.main,
     'augment': kevlar.augment.main,
     'assemble': kevlar.assemble.main,
     'mutate': kevlar.mutate.main,
@@ -51,7 +49,6 @@ subparser_funcs = {
     'effcount': effcount.subparser,
     'novel': novel.subparser,
     'filter': filter.subparser,
-    'reaugment': reaugment.subparser,
     'augment': augment.subparser,
     'assemble': assemble.subparser,
     'mutate': mutate.subparser,