+++ kevlar/tests/test_augment.py
@@ -13,7 +13,7 @@ import kevlar
 from kevlar.tests import data_file
 
 
-def test_augment_snorkel():
+def test_augment_contigs():
     augfh = kevlar.open(data_file('snorkel.augfastq'), 'r')
     augreads = kevlar.parse_augmented_fastx(augfh)
     nakedseq = screed.open(data_file('snorkel-contig.fasta'))
@@ -24,6 +24,20 @@ def test_augment_snorkel():
 
     offsets = [k.offset for k in augseqs[0].ikmers]
     assert offsets == [17, 20, 22]
+
+
+def test_augment_reads(capsys):
+    arglist = [
+        'augment',
+        data_file('reaugment.augfastq'),
+        data_file('reaugment.fq')
+    ]
+    args = kevlar.cli.parser().parse_args(arglist)
+    kevlar.augment.main(args)
+
+    out, err = capsys.readouterr()
+    testout = open(data_file('reaugment.out'), 'r').read()
+    assert out == testout
 
 
 def test_augment_cli(capsys):
@@ -41,16 +55,3 @@ AGGTCTTCGATGCTAGCATTTTTACGACAGACAAAAACAAGATTACATTCCAAAATACATACCGCGCC
                  ATTTTTACGAC          8 0 0#
                     TTTACGACAGA          11 0 0#
                       TACGACAGACA          9 0 0#"""
-
-
-def test_reaugment_basic(capsys):
-    args = kevlar.cli.parser().parse_args([
-        'reaugment',
-        data_file('reaugment.augfastq'),
-        data_file('reaugment.fq')
-    ])
-    kevlar.reaugment.main(args)
-
-    out, err = capsys.readouterr()
-    testout = open(data_file('reaugment.out'), 'r').read()
-    assert out == testout